+++ bin/node/src/commands/bundled.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, path::PathBuf};
+use std::{collections::HashMap, path::PathBuf, time::Duration};
 
 use anyhow::Context;
 use miden_node_block_producer::server::BlockProducer;
@@ -9,7 +9,8 @@ use tokio::{net::TcpListener, task::JoinSet};
 use url::Url;
 
 use super::{
-    ENV_BATCH_PROVER_URL, ENV_BLOCK_PROVER_URL, ENV_DATA_DIRECTORY, ENV_ENABLE_OTEL, ENV_RPC_URL,
+    DEFAULT_BATCH_INTERVAL_MS, DEFAULT_BLOCK_INTERVAL_MS, ENV_BATCH_PROVER_URL,
+    ENV_BLOCK_PROVER_URL, ENV_DATA_DIRECTORY, ENV_ENABLE_OTEL, ENV_RPC_URL, parse_duration_ms,
 };
 
 #[derive(clap::Subcommand)]
@@ -67,6 +68,24 @@ pub enum BundledCommand {
         /// OpenTelemetry documentation. See our operator manual for further details.
         #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "bool")]
         open_telemetry: bool,
+
+        /// Interval at which to produce blocks in milliseconds.
+        #[arg(
+            long = "block.interval",
+            default_value = DEFAULT_BLOCK_INTERVAL_MS,
+            value_parser = parse_duration_ms,
+            value_name = "MILLISECONDS"
+        )]
+        block_interval: Duration,
+
+        /// Interval at which to procude batches in milliseconds.
+        #[arg(
+            long = "batch.interval",
+            default_value = DEFAULT_BATCH_INTERVAL_MS,
+            value_parser = parse_duration_ms,
+            value_name = "MILLISECONDS"
+        )]
+        batch_interval: Duration,
     },
 }
 
@@ -95,7 +114,19 @@ impl BundledCommand {
                 block_prover_url,
                 // Note: open-telemetry is handled in main.
                 open_telemetry: _,
-            } => Self::start(rpc_url, data_directory, batch_prover_url, block_prover_url).await,
+                block_interval,
+                batch_interval,
+            } => {
+                Self::start(
+                    rpc_url,
+                    data_directory,
+                    batch_prover_url,
+                    block_prover_url,
+                    batch_interval,
+                    block_interval,
+                )
+                .await
+            },
         }
     }
 
@@ -104,6 +135,8 @@ impl BundledCommand {
         data_directory: PathBuf,
         batch_prover_url: Option<Url>,
         block_prover_url: Option<Url>,
+        batch_interval: Duration,
+        block_interval: Duration,
     ) -> anyhow::Result<()> {
         // Start listening on all gRPC urls so that inter-component connections can be created
         // before each component is fully started up.
@@ -140,6 +173,8 @@ impl BundledCommand {
             store_address,
             batch_prover_url,
             block_prover_url,
+            batch_interval,
+            block_interval,
         )
         .await
         .context("Loading block-producer")?;